+++ yapw/util.py
@@ -12,11 +12,11 @@ except ImportError:
 
 def json_dumps(message):
     """
-    Serializes the message to JSON formatted bytes.
+    Serializes a Python object to JSON formatted bytes.
 
     Uses `orjson <https://pypi.org/project/orjson/>`__ if available.
 
-    :param message: a JSON-serializable message
+    :param message: a Python object
     :returns: JSON formatted bytes
     :rtype: bytes
     """
@@ -25,11 +25,25 @@ def json_dumps(message):
     return orjson.dumps(message)
 
 
+def default_encode(message, content_type):
+    """
+    If the content type is "application/json", serializes the Python object to JSON formatted bytes. Otherwise, returns
+    the Python object (which might be bytes already).
+
+    :param message: a Python object
+    :returns: bytes
+    :rtype: bytes
+    """
+    if content_type == "application/json":
+        return json_dumps(message)
+    return message
+
+
 def basic_publish_kwargs(state, message, routing_key):
     """
     Prepares keyword arguments for ``basic_publish``.
 
-    :param state: an object with the attributes ``format_routing_key``, ``exchange``, ``encoder``, ``content_type`` and
+    :param state: an object with the attributes ``format_routing_key``, ``exchange``, ``encode``, ``content_type`` and
                   ``delivery_mode``
     :param message: a JSON-serializable message
     :param str routing_key: the routing key
@@ -38,8 +52,8 @@ def basic_publish_kwargs(state, message, routing_key):
     """
     formatted = state.format_routing_key(routing_key)
 
-    body = state.encoder(message)
-    properties = pika.BasicProperties(delivery_mode=state.delivery_mode, content_type=state.content_type)
+    body = state.encode(message, state.content_type)
+    properties = pika.BasicProperties(content_type=state.content_type, delivery_mode=state.delivery_mode)
 
     return {"exchange": state.exchange, "routing_key": formatted, "body": body, "properties": properties}